@@ -11,28 +11,28 @@
 		"lint": "eslint ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-static": "^3.0.5",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@tailwindcss/typography": "^0.5.14",
-		"@types/eslint": "^9.6.0",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/adapter-static": "^3.0.6",
+		"@sveltejs/kit": "^2.7.4",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@tailwindcss/typography": "^0.5.15",
+		"@types/eslint": "^9.6.1",
 		"autoprefixer": "^10.4.20",
 		"bits-ui": "^0.21.16",
-		"embla-carousel-svelte": "^8.3.0",
-		"eslint": "^9.0.0",
-		"eslint-plugin-svelte": "^2.36.0",
-		"globals": "^15.0.0",
+		"embla-carousel-svelte": "^8.3.1",
+		"eslint": "^9.14.0",
+		"eslint-plugin-svelte": "^2.46.0",
+		"globals": "^15.11.0",
 		"paneforge": "^0.0.6",
 		"postcss": "^8.4.47",
-		"svelte": "^4.2.7",
-		"svelte-check": "^4.0.0",
+		"svelte": "^5.1.9",
+		"svelte-check": "^4.0.5",
 		"svelte-radix": "^1.1.1",
-		"tailwindcss": "^3.4.9",
-		"typescript": "^5.0.0",
-		"typescript-eslint": "^8.0.0",
+		"tailwindcss": "^3.4.14",
+		"typescript": "^5.6.3",
+		"typescript-eslint": "^8.12.2",
 		"vaul-svelte": "^0.3.2",
-		"vite": "^5.0.3"
+		"vite": "^5.4.10"
 	},
 	"type": "module",
 	"dependencies": {
@@ -51,7 +51,7 @@
 		"clsx": "^2.1.1",
 		"localforage": "^1.10.0",
 		"svelte-gestures": "^5.0.6",
-		"tailwind-merge": "^2.5.3",
+		"tailwind-merge": "^2.5.4",
 		"tailwind-variants": "^0.2.1"
 	}
 }
