@@ -5,22 +5,34 @@
 
 	type $$Props = AccordionPrimitive.ContentProps;
 
-	let className: $$Props["class"] = undefined;
-	export let transition: $$Props["transition"] = slide;
-	export let transitionConfig: $$Props["transitionConfig"] = {
-		duration: 200,
-	};
+	interface Props {
+		class?: $$Props["class"];
+		transition?: $$Props["transition"];
+		transitionConfig?: $$Props["transitionConfig"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
 
-	export { className as class };
+	let {
+		class: className = undefined,
+		transition = slide,
+		transitionConfig = {
+		duration: 200,
+	},
+		children,
+		...rest
+	}: Props = $props();
+
+	
 </script>
 
 <AccordionPrimitive.Content
 	class={cn("overflow-hidden text-sm", className)}
 	{transition}
 	{transitionConfig}
-	{...$$restProps}
+	{...rest}
 >
 	<div class="pb-4 pt-0">
-		<slot />
+		{@render children?.()}
 	</div>
 </AccordionPrimitive.Content>
