@@ -4,11 +4,17 @@
 
 	type $$Props = AccordionPrimitive.ItemProps;
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
-	export let value: $$Props["value"];
+	
+	interface Props {
+		class?: $$Props["class"];
+		value: $$Props["value"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, value, children, ...rest }: Props = $props();
 </script>
 
-<AccordionPrimitive.Item {value} class={cn("border-b", className)} {...$$restProps}>
-	<slot />
+<AccordionPrimitive.Item {value} class={cn("border-b", className)} {...rest}>
+	{@render children?.()}
 </AccordionPrimitive.Item>
