@@ -6,9 +6,15 @@
 	type $$Props = AccordionPrimitive.TriggerProps;
 	type $$Events = AccordionPrimitive.TriggerEvents;
 
-	let className: $$Props["class"] = undefined;
-	export let level: AccordionPrimitive.HeaderProps["level"] = 3;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		level?: AccordionPrimitive.HeaderProps["level"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, level = 3, children, ...rest }: Props = $props();
+	
 </script>
 
 <AccordionPrimitive.Header {level} class="flex">
@@ -17,10 +23,10 @@
 			"flex flex-1 items-center justify-between py-4 text-sm font-medium transition-all hover:underline [&[data-state=open]>svg]:rotate-180",
 			className
 		)}
-		{...$$restProps}
+		{...rest}
 		on:click
 	>
-		<slot />
+		{@render children?.()}
 		<ChevronDown
 			class="text-muted-foreground h-4 w-4 shrink-0 transition-transform duration-200"
 		/>
