@@ -4,9 +4,15 @@
 
 	type $$Props = CalendarPrimitive.CellProps;
 
-	export let date: $$Props["date"];
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		date: $$Props["date"];
+		class?: $$Props["class"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { date, class: className = undefined, children, ...rest }: Props = $props();
+	
 </script>
 
 <CalendarPrimitive.Cell
@@ -15,7 +21,7 @@
 		"[&:has([data-selected])]:bg-accent [&:has([data-selected][data-outside-month])]:bg-accent/50 relative p-0 text-center text-sm focus-within:relative focus-within:z-20 [&:has([data-selected])]:rounded-md",
 		className
 	)}
-	{...$$restProps}
+	{...rest}
 >
-	<slot />
+	{@render children?.()}
 </CalendarPrimitive.Cell>
