@@ -7,8 +7,14 @@
 	type $$Props = CalendarPrimitive.PrevButtonProps;
 	type $$Events = CalendarPrimitive.PrevButtonEvents;
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		children?: import('svelte').Snippet<[any]>;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, children, ...rest }: Props = $props();
+	
 </script>
 
 <CalendarPrimitive.PrevButton
@@ -18,10 +24,12 @@
 		"h-7 w-7 bg-transparent p-0 opacity-50 hover:opacity-100",
 		className
 	)}
-	{...$$restProps}
-	let:builder
+	{...rest}
+	
 >
-	<slot {builder}>
-		<ChevronLeft class="h-4 w-4" />
-	</slot>
+	{#snippet children({ builder })}
+		{#if children}{@render children({ builder, })}{:else}
+			<ChevronLeft class="h-4 w-4" />
+		{/if}
+	{/snippet}
 </CalendarPrimitive.PrevButton>
