@@ -5,15 +5,21 @@
 	type $$Props = HTMLAttributes<HTMLDivElement> & {
 		el?: HTMLDivElement;
 	};
-	export let el: $$Props["el"] = undefined;
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		el?: $$Props["el"];
+		class?: $$Props["class"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { el = $bindable(undefined), class: className = undefined, children, ...rest }: Props = $props();
+	
 </script>
 
 <div
 	bind:this={el}
 	class={cn("grid gap-1.5 p-4 text-center sm:text-left", className)}
-	{...$$restProps}
+	{...rest}
 >
-	<slot />
+	{@render children?.()}
 </div>
