@@ -2,11 +2,23 @@
 	import { Drawer as DrawerPrimitive } from "vaul-svelte";
 
 	type $$Props = DrawerPrimitive.Props;
-	export let shouldScaleBackground: $$Props["shouldScaleBackground"] = true;
-	export let open: $$Props["open"] = false;
-	export let activeSnapPoint: $$Props["activeSnapPoint"] = undefined;
+	interface Props {
+		shouldScaleBackground?: $$Props["shouldScaleBackground"];
+		open?: $$Props["open"];
+		activeSnapPoint?: $$Props["activeSnapPoint"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let {
+		shouldScaleBackground = true,
+		open = $bindable(false),
+		activeSnapPoint = $bindable(undefined),
+		children,
+		...rest
+	}: Props = $props();
 </script>
 
-<DrawerPrimitive.Root {shouldScaleBackground} bind:open bind:activeSnapPoint {...$$restProps}>
-	<slot />
+<DrawerPrimitive.Root {shouldScaleBackground} bind:open bind:activeSnapPoint {...rest}>
+	{@render children?.()}
 </DrawerPrimitive.Root>
