@@ -6,9 +6,14 @@
 		orientation?: "vertical" | "horizontal";
 	};
 
-	let className: $$Props["class"] = undefined;
-	export let orientation: $$Props["orientation"] = "vertical";
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		orientation?: $$Props["orientation"];
+		children?: import('svelte').Snippet;
+	}
+
+	let { class: className = undefined, orientation = "vertical", children }: Props = $props();
+	
 </script>
 
 <ScrollAreaPrimitive.Scrollbar
@@ -20,7 +25,7 @@
 		className
 	)}
 >
-	<slot />
+	{@render children?.()}
 	<ScrollAreaPrimitive.Thumb
 		class={cn("bg-border relative rounded-full", orientation === "vertical" && "flex-1")}
 	/>
