@@ -4,10 +4,15 @@
 
 	type $$Props = SeparatorPrimitive.Props;
 
-	let className: $$Props["class"] = undefined;
-	export let orientation: $$Props["orientation"] = "horizontal";
-	export let decorative: $$Props["decorative"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		orientation?: $$Props["orientation"];
+		decorative?: $$Props["decorative"];
+		[key: string]: any
+	}
+
+	let { class: className = undefined, orientation = "horizontal", decorative = undefined, ...rest }: Props = $props();
+	
 </script>
 
 <SeparatorPrimitive.Root
@@ -18,5 +23,5 @@
 	)}
 	{orientation}
 	{decorative}
-	{...$$restProps}
+	{...rest}
 />
