@@ -4,10 +4,16 @@
 
 	type $$Props = HTMLAttributes<HTMLTableCaptionElement>;
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, children, ...rest }: Props = $props();
+	
 </script>
 
-<caption class={cn("text-muted-foreground mt-4 text-sm", className)} {...$$restProps}>
-	<slot />
+<caption class={cn("text-muted-foreground mt-4 text-sm", className)} {...rest}>
+	{@render children?.()}
 </caption>
