@@ -4,12 +4,18 @@
 
 	type $$Props = HTMLTableAttributes;
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, children, ...rest }: Props = $props();
+	
 </script>
 
 <div class="relative w-full overflow-auto">
-	<table class={cn("w-full caption-bottom text-sm", className)} {...$$restProps}>
-		<slot />
+	<table class={cn("w-full caption-bottom text-sm", className)} {...rest}>
+		{@render children?.()}
 	</table>
 </div>
