@@ -1,5 +1,12 @@
 <script>
 	import '../app.css';
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
 </script>
 
-<slot></slot>
+{@render children?.()}
