@@ -6,12 +6,23 @@
 	type $$Props = CalendarPrimitive.Props;
 	type $$Events = CalendarPrimitive.Events;
 
-	export let value: $$Props["value"] = undefined;
-	export let placeholder: $$Props["placeholder"] = undefined;
-	export let weekdayFormat: $$Props["weekdayFormat"] = "short";
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		value?: $$Props["value"];
+		placeholder?: $$Props["placeholder"];
+		weekdayFormat?: $$Props["weekdayFormat"];
+		class?: $$Props["class"];
+		[key: string]: any
+	}
+
+	let {
+		value = $bindable(undefined),
+		placeholder = $bindable(undefined),
+		weekdayFormat = "short",
+		class: className = undefined,
+		...rest
+	}: Props = $props();
+	
 </script>
 
 <div>
@@ -20,44 +31,46 @@
 		bind:placeholder
 		{weekdayFormat}
 		class={cn("p-3", className)}
-		{...$$restProps}
+		{...rest}
 		on:keydown
-		let:months
-		let:weekdays
+		
+		
 	>
-		<Calendar.Header>
-			<Calendar.PrevButton />
-			<Calendar.Heading />
-			<Calendar.NextButton />
-		</Calendar.Header>
-		<Calendar.Months>
-			{#each months as month}
-				<Calendar.Grid>
-					<Calendar.GridBody class="w-fit items-center justify-center">
-						{#each month.weeks as weekDates}
-							<Calendar.GridRow class="mt-2 justify-center">
-								{#each weekDates as date}
-									<Calendar.Cell {date}>
-										<Calendar.Day {date} month={month.value} />
-									</Calendar.Cell>
+		{#snippet children({ months, weekdays })}
+				<Calendar.Header>
+				<Calendar.PrevButton />
+				<Calendar.Heading />
+				<Calendar.NextButton />
+			</Calendar.Header>
+			<Calendar.Months>
+				{#each months as month}
+					<Calendar.Grid>
+						<Calendar.GridBody class="w-fit items-center justify-center">
+							{#each month.weeks as weekDates}
+								<Calendar.GridRow class="mt-2 justify-center">
+									{#each weekDates as date}
+										<Calendar.Cell {date}>
+											<Calendar.Day {date} month={month.value} />
+										</Calendar.Cell>
+									{/each}
+								</Calendar.GridRow>
+							{/each}
+						</Calendar.GridBody>
+
+						<Calendar.GridHead>
+							<Calendar.GridRow class="flex w-full items-center justify-center">
+								{#each weekdays as weekday}
+									<Calendar.HeadCell>
+										{weekday.slice(0, 2)}
+									</Calendar.HeadCell>
 								{/each}
 							</Calendar.GridRow>
-						{/each}
-					</Calendar.GridBody>
-
-					<Calendar.GridHead>
-						<Calendar.GridRow class="flex w-full items-center justify-center">
-							{#each weekdays as weekday}
-								<Calendar.HeadCell>
-									{weekday.slice(0, 2)}
-								</Calendar.HeadCell>
-							{/each}
-						</Calendar.GridRow>
-					</Calendar.GridHead>
-				</Calendar.Grid>
-			{/each}
-		</Calendar.Months>
-	</CalendarPrimitive.Root>
+						</Calendar.GridHead>
+					</Calendar.Grid>
+				{/each}
+			</Calendar.Months>
+					{/snippet}
+		</CalendarPrimitive.Root>
 </div>
 
 <style>
